+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 VictoryCore Changelog
 =====================
 
+## 13.0.1 (2017-01-30)
+
+- [193](https://github.com/FormidableLabs/victory-core/pull/193)
+  - Removes the default `standalone: true` from VictoryContainer so that parents can set the value
+
+## 13.0.0 (2017-01-30)
+
+- [189](https://github.com/FormidableLabs/victory-core/pull/189) and [191](https://github.com/FormidableLabs/victory-core/pull/190)
+  - Adds `VictoryLegend` component
+- [190](https://github.com/FormidableLabs/victory-core/pull/190)
+  - Allows `VictoryContainer` to render either `<g>` or `<svg>` depending on the value of the standalone prop
+  - Passes a timer down in context for `VictorySharedEvents`
+  - Event handlers have access to the context they are being called from via an argument
+    - _i.e._ `onClick: (event, targetProps, eventKey, context) => {...}`
+  - Enhances `addEvents` so that evented components can pick up "parentControllerProps" from parent state
+    - useful for `VictoryZoomContainer`
+  - Adds the ability to define callbacks in the events prop that will be called after `setState`
+    - useful for allowing `VictoryZoomContainer` to call methods like `resumeAnimation`
+
 ## 12.0.2 (2017-01-14)
 
 - [188](https://github.com/FormidableLabs/victory-core/pull/188)